% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eve_Fl.R
\name{eve_Fl}
\alias{eve_Fl}
\title{Process FLAIR Neuroimages and Register to EVE Brain Template}
\usage{
eve_Fl(fpath, outpath, fsl_path, fsl_outputtype = "NIFTI_GZ")
}
\arguments{
\item{fpath}{Character string specifying the path to the FLAIR image file. The file should be in NIFTI file format (.nii.gz).}

\item{outpath}{Character string specifying the output directory where the processed data is saved.}

\item{fsl_path}{Character string specifying the path to the FSL software on the system.}

\item{fsl_outputtype}{Character string specifying the type of output file format for FSL; defaults to "NIFTI_GZ".}
}
\value{
Returns a list containing three elements: \code{intensities}, \code{tissues}, and \code{brain_volume_cm3}.
Each element corresponds to the array of intensities, the segmented tissue data, and the calculated brain volumes,
respectively. The function also saves these results as an .Rdata file at the specified output path.
}
\description{
This function processes FLAIR (Fluid-Attenuated Inversion Recovery) neuroimages for better lesion detection,
especially in the periventricular area by suppressing the CSF signal. It involves steps such as reading the image,
reorienting, bias correction using N4, brain extraction, registration to the EVE template, and tissue segmentation.
It finally calculates the intracranial volume and outputs the data.
}
\details{
The function uses specific FSL tools for image processing steps such as reorientation to standard space,
bias correction with N4 method from ANTsR, and brain extraction using a robust method from \code{extrantsr}.
Segmentation into different tissue types (CSF, grey matter, and white matter) is performed using FSL's FAST tool.
Volumes are calculated based on the segmented tissues.
}
\examples{
\dontrun{
   eve_Fl("path/to/your/flair/image.nii.gz",
         "path/to/output/",
         "/usr/local/fsl",
         "NIFTI_GZ")
}

}
