% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2_eve.R
\name{map2_eve}
\alias{map2_eve}
\title{Plot Mask on EVE Template using Python and Nilearn}
\usage{
map2_eve(mask_img_path, cmap = "bwr_r", alpha = 1, save_path = NULL, ...)
}
\arguments{
\item{mask_img_path}{A string representing the file path to the mask NIfTI file.}

\item{cmap}{The colormap to use. Either a string (name of a matplotlib colormap) or a matplotlib colormap object. Default is 'bwr_r'.}

\item{alpha}{Transparency level for the overlay. Default is 1.}

\item{save_path}{A string representing the file path where the plot image should be saved (e.g., "output.png"). If NULL, the plot will be shown interactively instead of saved. Default is NULL.}

\item{...}{Additional arguments to customize the anatomical plot. These arguments are passed
directly to the Python function \code{nilearn.plotting.plot_anat}. You can specify parameters
such as \code{title}, \code{display_mode}, \code{cut_coords}, \code{dim}, etc.
For more details on available options, refer to the official documentation at:
https://nilearn.github.io/stable/modules/generated/nilearn.plotting.plot_anat.html}
}
\value{
None
}
\description{
This function plots a mask image on the EVE template using Python's Nilearn library and optionally saves the plot as an image.
It ensures the required Python libraries are installed, and it handles the appropriate Conda environment setup.
}
\details{
The function first detects the system architecture (ARM or x86) and ensures the appropriate Conda environment is set up (Miniforge for ARM or Miniconda for x86).
It then checks and installs the required Python libraries: \code{nilearn}, \code{nibabel}, and \code{matplotlib}.
The function loads the Python environment and calls the \code{plot_mask_on_eve} function, which plots the mask on the EVE template. If \code{save_path} is provided, the plot is saved as an image at the specified location.
}
\examples{
\dontrun{
map2_eve(
  mask_img_path = "/path/to/mask_nifti_GM_Volume.nii.gz",
  save_path = "/path/to/save_output.png",
  cmap = "bwr_r",
  alpha = 0.8,
  title = "Mask on EVE Template"
)
}

}
