% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartitionPipeline.R
\name{PartitionPipeline}
\alias{PartitionPipeline}
\title{Partition Pipeline for Image Analysis}
\description{
This R6 class is designed to streamline the processing pipeline for image analysis,
including steps from initial processing to combining independent variables with
reduced variables by tissue type by ROI.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PartitionPipeline-new}{\code{PartitionPipeline$new()}}
\item \href{#method-PartitionPipeline-iproc}{\code{PartitionPipeline$iproc()}}
\item \href{#method-PartitionPipeline-supparfun}{\code{PartitionPipeline$supparfun()}}
\item \href{#method-PartitionPipeline-map_suppar_roi}{\code{PartitionPipeline$map_suppar_roi()}}
\item \href{#method-PartitionPipeline-parfun}{\code{PartitionPipeline$parfun()}}
\item \href{#method-PartitionPipeline-tissue_segment}{\code{PartitionPipeline$tissue_segment()}}
\item \href{#method-PartitionPipeline-Cmb_tissue_type}{\code{PartitionPipeline$Cmb_tissue_type()}}
\item \href{#method-PartitionPipeline-process_indep_variables}{\code{PartitionPipeline$process_indep_variables()}}
\item \href{#method-PartitionPipeline-Cmb_indep_with_dep}{\code{PartitionPipeline$Cmb_indep_with_dep()}}
\item \href{#method-PartitionPipeline-clone}{\code{PartitionPipeline$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-new"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$new(
  tind = NULL,
  nfl = NULL,
  main_dir = NULL,
  tissue_type = NULL,
  outp_volume = TRUE,
  ICC_thresh_vec = NULL,
  suppar_thresh_vec = seq(0.7, 1, 0.01),
  B = 2000,
  roi = NULL,
  num_cores = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-iproc"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-iproc}{}}}
\subsection{Method \code{iproc()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$iproc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-supparfun"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-supparfun}{}}}
\subsection{Method \code{supparfun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$supparfun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-map_suppar_roi"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-map_suppar_roi}{}}}
\subsection{Method \code{map_suppar_roi()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$map_suppar_roi()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-parfun"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-parfun}{}}}
\subsection{Method \code{parfun()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$parfun()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-tissue_segment"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-tissue_segment}{}}}
\subsection{Method \code{tissue_segment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$tissue_segment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-Cmb_tissue_type"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-Cmb_tissue_type}{}}}
\subsection{Method \code{Cmb_tissue_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$Cmb_tissue_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-process_indep_variables"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-process_indep_variables}{}}}
\subsection{Method \code{process_indep_variables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$process_indep_variables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-Cmb_indep_with_dep"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-Cmb_indep_with_dep}{}}}
\subsection{Method \code{Cmb_indep_with_dep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$Cmb_indep_with_dep()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PartitionPipeline-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PartitionPipeline-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PartitionPipeline$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
