% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_partition_pipeline.R
\name{run_partition_pipeline}
\alias{run_partition_pipeline}
\title{Run Partition Pipeline on Neuroimaging Data}
\usage{
run_partition_pipeline(
  tind,
  nfl,
  main_dir,
  tissue_type,
  ICC_thresh_vec,
  num_cores,
  suppar_thresh_vec = seq(0.7, 1, 0.01),
  B = 2000,
  outp_volume = TRUE
)
}
\arguments{
\item{tind}{Index or identifier for the type of tissue under analysis.}

\item{nfl}{List of file names (full paths) that need to be processed.}

\item{main_dir}{Main directory where outputs and intermediate results will be saved.}

\item{tissue_type}{Type of tissue for segmentation.}

\item{ICC_thresh_vec}{A vector of threshold values for Intraclass Correlation Coefficient used in the Partition Algorithm.}

\item{num_cores}{Number of cores to use for parallel processing.}

\item{suppar_thresh_vec}{Optional; a sequence of threshold values used in Super Partitioning.
Default is a sequence from 0.7 to 1 by 0.01.}

\item{B}{Optional; the maximum size of modules to be considered in partitioning. Default is 2000.}

\item{outp_volume}{Optional; a logical indicating whether volume outputs should be generated. Default is TRUE.}
}
\value{
The function does not return a value but will output results directly to
the specified \code{main_dir} as side effects of the processing steps.
}
\description{
This function initializes and executes a partitioning pipeline designed for
processing neuroimaging data. It handles tasks such as image processing,
super partition analysis, mapping, and combining of data based on specified
thresholds and parameters.
}
\details{
The function configures and runs a series of operations that are typical in
neuroimage analysis, especially focusing on ROI-based transformations. Each step
of the pipeline, from initial image processing (\code{iproc}) to the final combination
of independent variables with reduced variables (\code{Cmb_indep_with_dep}), is executed
in sequence. Adjustments to the pipeline's behavior can be made by changing the
function parameters, allowing for custom analysis flows.
}
\examples{
run_partition_pipeline(tind = 5,
                       nfl = list.files('/path/to/data', full.names = TRUE),
                       main_dir = "/path/to/output",
                       tissue_type = 2,
                       ICC_thresh_vec = c(0.8, 0.9),
                       num_cores = 4)

}
